# src/services/requests/IUpdateUserRequest.ts
export interface IUpdateUserRequest {
  nickName: string;
  gender: string;
  description: string;
  city: string;
  email: string;
  mobile: string;
  wechatId: string;
  company: string;
  avatarImage: string;
  ageTag: string[];
}