# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"IResponse": {
			"properties": {
				"code": {
					"type": "string"
				},
				"data": {
					"additionalProperties": true,
					"type": "object"
				}
			},
			"required": [
				"code",
				"data"
			],
			"type": "object",
			"additionalProperties": true
		},
		"IErrorResponse": {
			"properties": {
				"status": {
					"type": "number",
					"format": "double"
				},
				"code": {
					"type": "string"
				},
				"data": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"status",
				"code"
			],
			"type": "object",
			"additionalProperties": true
		},
		"IOauthLoginRequest": {
			"properties": {
				"appName": {
					"type": "string"
				},
				"code": {
					"type": "string"
				},
				"type": {
					"type": "string"
				},
				"nickName": {
					"type": "string"
				},
				"avatarUrl": {
					"type": "string"
				},
				"gender": {
					"type": "string"
				},
				"encryptedData": {
					"type": "string"
				},
				"iv": {
					"type": "string"
				},
				"country": {
					"type": "string"
				},
				"province": {
					"type": "string"
				},
				"city": {
					"type": "string"
				},
				"language": {
					"type": "string"
				},
				"description": {
					"type": "string"
				}
			},
			"required": [
				"appName",
				"code",
				"type",
				"nickName",
				"avatarUrl",
				"gender",
				"encryptedData",
				"iv",
				"country",
				"province",
				"city",
				"language",
				"description"
			],
			"type": "object",
			"additionalProperties": true
		},
		"IUserResponse": {
			"properties": {
				"code": {
					"type": "string"
				},
				"data": {
					"properties": {},
					"type": "object"
				}
			},
			"required": [
				"code",
				"data"
			],
			"type": "object",
			"additionalProperties": true
		},
		"IUpdateUserRequest": {
			"properties": {
				"nickName": {
					"type": "string"
				},
				"gender": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"city": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"mobile": {
					"type": "string"
				},
				"wechatId": {
					"type": "string"
				},
				"company": {
					"type": "string"
				},
				"avatarImage": {
					"type": "string"
				},
				"ageTag": {
					"items": {
						"type": "string"
					},
					"type": "array"
				}
			},
			"required": [
				"nickName",
				"gender",
				"description",
				"city",
				"email",
				"mobile",
				"wechatId",
				"company",
				"avatarImage",
				"ageTag"
			],
			"type": "object",
			"additionalProperties": true
		}
	},
	"info": {
		"title": "jubenfan-server",
		"version": "3.0.0",
		"description": "API description can go here",
		"license": {
			"name": "MIT"
		}
	},
	"paths": {
		"/oauth/login": {
			"post": {
				"operationId": "socialLogin",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/IResponse"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"$ref": "#/definitions/IErrorResponse"
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IOauthLoginRequest"
						}
					}
				]
			}
		},
		"/profile": {
			"get": {
				"operationId": "getProfile",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/IUserResponse"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"$ref": "#/definitions/IErrorResponse"
						}
					}
				},
				"tags": [
					"profile"
				],
				"security": [
					{
						"access_token": []
					}
				],
				"parameters": []
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "getUserDetailById",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/IResponse"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"$ref": "#/definitions/IErrorResponse"
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"access_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "updateUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/IResponse"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"$ref": "#/definitions/IErrorResponse"
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"access_token": [
							"user:updateUser"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IUpdateUserRequest"
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header",
			"template": "Bearer {apiKey}"
		}
	},
	"host": "localhost:3000",
	"schemes": [
		"http",
		"https"
	]
}